refactor: consolidate resolveTurnForGame integration tests (#3842)
Add turn_resolver_test_harness, merge 11 partN_segmentM files into four
domain-grouped integration tests, and drop economy preview back-compat wrappers.
